fix(order): prevent multiple handouts of a blid (#443)
* fix(order): prevent multiple handouts of a blid

There was already a check to prevent a blid from being added to an order
if it already belongs to another order, but if the order had been
fulfilled (i.e. the blid handed out), the item could be handed out to
another person without first being handed in. This commit fixes that by
checking whether some blid in an order is not handed in. In that case,
it returns an error with code 801 (409 in the HTTP response).

There are some customeritems which do not have blids, which means we
cannot prevent those items from being handed out again. However, some
research indicates that these are legacy items of which there are
exceedingly few which are not handed in (~55 at time of writing).

* chore: refactor fix-multiple-handout