feat(match): update endpoint to use new matching
Uses variants in the Match schema with a key "_variant" and optional
keys for those keys which are only present in some variants. Those have
default value undefined to not show up in the DB browser unnecessarily.
Because of the Mongoose schema, no two variants can have different types
under the same key.