Create new match endpoints (#447)
* refactor(BlCollection): extract BlCollection-names to enum

* refactor(match): Disambiguate and update different Match models

There are two kinds of Match model: those MatchFinder uses, which
contains only the necessary information, and the model which is stored
in the database. Those had confusing names and structures, which is
somewhat improved with this commit.

* feat(match): update endpoint to use new matching

Uses variants in the Match schema with a key "_variant" and optional
keys for those keys which are only present in some variants. Those have
default value undefined to not show up in the DB browser unnecessarily.
Because of the Mongoose schema, no two variants can have different types
under the same key.

* chore: run prettier

* chore(bl-model): bump bl-model with match type changes

* refactor(match): improve readability and error handling

- Extract common elements of getMatchableSenders and -Receivers
- Rename and document some helper functions
- Add super to generate endpoint
- Allow nullability of Match.meetingInfo.date in mongoose schema
- Be stricter when verifying MatchGenerateOperation input data
- Use the correct error code for bad request