feat: add digital signatures
Adds support for digital signatures through the new /signatures
endpoints, as well as the pendingSignature field of Order. This will
require a customer to have had signed the loaning agreement prior to
handout. This only applies to loans, not partly-payments or matches. The
agreement has an expiration date of 4 years before needing to be
re-signed, or until the customer becomes of age. Orders which had a
valid signature at order time can still be handed out after the
signature expires. The signature images are stored as image binaries in
the database, but everywhere else they are serialized to base64 for
ease of transfer. The images are converted to webp and compressed
pretty heavily before storage.

This sadly also includes a bit of a refactor:
* mongoDb.blStorageHandler now returns POJOs instead of thick
  Mongoose Document types. This fixes spreading using the ... operator,
  and generally makes our TypeScript types more accurately reflect the
  actual runtime values of BlDocuments. This required some changes to
  the document transform operation as well.
* updated error printing in bl-error-handler to include non-BlError
  stack traces.
* better logging of unhandled rejections (non-awaited promises).
* upped request body size limit to receive uncompressed signatures.