fix: ignore user change, update reminders spring 2024 (#505)
* build(deps): bump bl-post-office for reminders spring 2024

* fix(mongoDb): omit user change instead of erroring

Some calls, like updating branch info, include the "user" field of a
BlDocument in their "update" call. If the user who created the document
is not the same as the one updating it, this would error even though the
updater has permission to update the document. At the same time, we
never want to change ownership of the document. Thus, we should omit the
user change in update calls, instead of erroring.