refactor(schemas): fix ObjectId/string type errors
Previously, the bl-model type for most references to other collections
was e.g. 'string | Order' if the reference was to the Order collection.
By changing this to 'string' only, we get better type safety, but
Mongoose now caused TS errors because the schema requires an ObjectId,
but our type requires a string. The new helper type 'ToSchema' alters
the types to change every 'string' to 'string | ObjectId', except if the
field is named "type", which is a reserved name and can only be string.