feat(match-transfer): use receiver branch
This fixes a few problems:
- When receiving a book, it was handed out using the sender's branch,
not the receiver's. This impacted which deadline was used, and which
branch the customeritem was registered on.
- Set movedFrom for match-receive orders to the receiver's original
order.
- Fix bug with movedTo/-From where the order which was moved from would
not get an update to its movedTo. This was caused by a type discrepancy
where a string item-ID was checked against an ObjectId, which always
failed. The root cause was a lack of recursion in the
MongooseSchemaCreator.transformObject-function, where objects nested in
arrays were not transformed.