match transfer changes (#534)
* refactor(match-transfer): extract helpers

* feat(match-transfer): use original order deadline

Did not remove deadlineOverride, so now the priority order goes
deadlineOverride > original receiver order deadline > branch rent
period.

* feat(match-transfer): use receiver branch

This fixes a few problems:
- When receiving a book, it was handed out using the sender's branch,
not the receiver's. This impacted which deadline was used, and which
branch the customeritem was registered on.
- Set movedFrom for match-receive orders to the receiver's original
order.
- Fix bug with movedTo/-From where the order which was moved from would
not get an update to its movedTo. This was caused by a type discrepancy
where a string item-ID was checked against an ObjectId, which always
failed. The root cause was a lack of recursion in the
MongooseSchemaCreator.transformObject-function, where objects nested in
arrays were not transformed.

* feat(match-transfer): attach customeritem to orders

* fix(DateService): fix flakiness in UTC-diff-zone

The test for betweenHours() failed between 00:00 and 02:00 in Norwegian
summer time (UTC+2) because the wrong day was used. Fixed by
constructing dates with the supplied timezone.

* fix(match-operation): remove unneccecary check

* fix(match-generate): add missing ObjectID conversion

---------

Co-authored-by: Adrian Andersen <adrianandersen@protonmail.com>